hps_accel/stream: split actor into own file
Splits BinaryCombinatorialActor into its own file, actor.py, in
preparation for future commits which will add new types of actors.
