+++ proj/hps_accel/gateware/stream/stream.py
@@ -13,10 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from migen.fhdl.module import Module
-from nmigen import Signal, Shape
+from nmigen import Shape
 from nmigen.hdl.rec import Record, DIR_FANIN, DIR_FANOUT
-from util import SimpleElaboratable
 
 
 class _Endpoint:
@@ -106,44 +104,3 @@ def flowcontrol_passthrough(sink, source):
         source.valid.eq(sink.valid),
         source.last.eq(sink.last),
     ]
-
-
-class BinaryCombinatorialActor(SimpleElaboratable):
-    """Base for a combinatorial binary actor.
-
-    Performs a combinatorial operation on a sink payload and routes it to a
-    source.
-
-    Parameters
-    ----------
-
-    input_type: The Shape or Layout for the sink
-
-    output_type: The Shape or Layout for the source.
-
-    Attributes:
-
-    sink: Sink(input_type), in
-      Sink for incoming data
-
-    source: Source(source_type), out
-      Source for outgoing data
-    """
-
-    def __init__(self, input_type, output_type):
-        self.input_type = input_type
-        self.output_type = output_type
-        self.sink = Sink(input_type)
-        self.source = Source(output_type)
-
-    def elab(self, m: Module):
-        m.d.comb += flowcontrol_passthrough(self.sink, self.source)
-        self.transform(m, self.sink.payload, self.source.payload)
-
-    def transform(self, m, input, output):
-        """Transforms input to output.
-
-        input: self.input_type, in
-        output: self.output_type, out
-        """
-        raise NotImplemented()
